requirements.txt
requests==2.31.0
google-api-python-client==2.108.0
moviepy==1.0.3
pyttsx3==2.90
python-dotenv==1.0.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.108.0
gTTS==2.3.1
Pillow==10.2.0
